v0.4.4 + docs

+++ application/GW2 Addon Manager/arcdps.cs
@@ -50,7 +50,7 @@ namespace GW2_Addon_Manager
             dynamic config_obj = configuration.getConfig();
             string game_path = config_obj.game_path;
             string bin64 = game_path + "\\" + config_obj.bin_folder + "\\";
-            if (config_obj.installed.arcdps != null)
+            if (config_obj.installed.arcdps != null && File.Exists(bin64 + config_obj.installed.arcdps))
                 File.Move(bin64 + config_obj.installed.arcdps, "Disabled Plugins\\arcdps.dll");
 
             config_obj.disabled.arcdps = true;
