+++ application/GW2 Addon Manager/arcdps_mechanics.cs
@@ -40,7 +40,7 @@ namespace GW2_Addon_Manager
             dynamic config_obj = configuration.getConfig();
             string game_path = config_obj.game_path;
             string bin64 = game_path + "\\" + config_obj.bin_folder + "\\";
-            if (config_obj.installed.arcdps_mechanics != null)
+            if (config_obj.installed.arcdps_mechanics != null && File.Exists(bin64 + config_obj.installed.arcdps_mechanics))
                 File.Move(bin64 + config_obj.installed.arcdps_mechanics, "Disabled Plugins\\" + dll_name);
 
             config_obj.disabled.arcdps_mechanics = true;
